% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily.R
\name{daily}
\alias{daily}
\title{Create derived (daily summary) measurements for multiple subjects}
\usage{
daily(data, subj = NULL, plotting = TRUE)
}
\arguments{
\item{data}{List of data_grid outcomes}

\item{subj}{Subject name}

\item{plotting}{If TRUE, a ggplot object is produced.}
}
\description{
Create derived (daily summary) measurements for multiple subjects
}
