% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_grid.R
\name{daily_grid}
\alias{daily_grid}
\title{Stretch derived (daily summary) measurements for multiple subjects (evaluate curves on a common grid)}
\usage{
daily_grid(data, ngrid = 200)
}
\arguments{
\item{data}{Long data frame of daily measure. Output of function 'daily'. Data include columns named Subj, Date, and Measure.}

\item{ngrid}{Number of grid. Default is 200.}
}
\description{
Stretch derived (daily summary) measurements for multiple subjects (evaluate curves on a common grid)
}
