% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_grid.R
\name{data_grid}
\alias{data_grid}
\title{Structure the observed data into an equally-spaced data frame}
\usage{
data_grid(data, variable, interval = 3, na.missing = FALSE,
  smoothing = TRUE)
}
\arguments{
\item{data}{Data in a matrix form. Data should contain a variable named 'timestamp' in a 'yyyy-mm-dd hh:mm:ss' format.}

\item{variable}{Name of variable in the data set you would like to re-structure}

\item{interval}{The interval of time in second. Default is 3 (i.e. The measurement is evaluated every three second.)}

\item{na.missing}{If TRUE, missing is considered as NA. If FALSE, the missing is considered as 0. Defulat is FALSE.}

\item{smoothing}{If TRUE, the curve is smooted first and evaluated at every interval minutes. Defulat is TRUE. If na.missing is TRUE, smoothing is automatically FALSE.}
}
\description{
Structure the observed data into an equally-spaced data frame
}
