% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPCA.R
\name{FPCA}
\alias{FPCA}
\title{Run FPCA (daily / all combined)}
\usage{
FPCA(data, nbasis = 12)
}
\arguments{
\item{data}{'data_grid' object}

\item{nbasis}{Number of basis. Defult is 12.}
}
\description{
Run FPCA (daily / all combined)
}
