% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPCA_plot.R
\name{FPCA_plot}
\alias{FPCA_plot}
\title{Plot FPCA (Overall)}
\usage{
FPCA_plot(fpca, npc = 2, obj = c("FPC", "EST"))
}
\arguments{
\item{fpca}{FPCA object: Specifically, output 'FPCA' of FPCA()}

\item{npc}{Number of FPCs to plot. Default is 2.}

\item{obj}{Objects for plotting. Default is FPCs. Possible options include FPCs and estimated curve.}
}
\description{
Plot FPCA (Overall)
}
