% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPCA_plot_day.R
\name{FPCA_plot_day}
\alias{FPCA_plot_day}
\title{Plot FPCA (Daily)}
\usage{
FPCA_plot_day(fpca, npc = 2, obj = c("FPC", "EST"))
}
\arguments{
\item{fpca}{FPCA object: Specifically, output 'FPCA_day' of FPCA()}

\item{npc}{Number of FPCs to plot. Default is 2. Option for FPC plot}

\item{obj}{Objects for plotting. Default is FPCs. Possible options include FPCs and estimated curve.}
}
\description{
Plot FPCA (Daily)
}
