% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_plot.R
\name{raw_plot}
\alias{raw_plot}
\title{Plot the daily raw trajectory per individual}
\usage{
raw_plot(data)
}
\arguments{
\item{data}{Data in a data.frame format; output of function 'data_grid'}
}
\description{
Plot the daily raw trajectory per individual
}
